Refactoing: Reorganized Large Classes into smaller files
