+++ run.py
@@ -1,10 +1,4 @@
-from ui.ui import create_main_window
-from main import start_application
-
-def run():
-    root = create_main_window()
-    start_application(root)
-    root.mainloop()
+from main import main  # Import the main function directly
 
 if __name__ == "__main__":
-    run()
+    main()
